Extract StdioProxy and MCPService from the CLI entry point
